JAMES-3788 Prevent NPE upon data-races when receiving PROXY command
SMTP protocol handlers runs in a dedicated executor
while HAProxy was running directly on the event loop.

There was thus a chance that the processing of the
HAProxy message to be processed before channel
information be correctly setted up resulting on a NPE
eventually closing the whole connection.

We move HA Proxy message handling on the same thread
that connection information set up which prevents
this data race.
