+++ protocols/netty/src/main/java/org/apache/james/protocols/netty/AbstractChannelPipelineFactory.java
@@ -68,7 +68,6 @@ public abstract class AbstractChannelPipelineFactory<C extends SocketChannel> ex
 
         if (proxyRequired) {
             pipeline.addLast(HandlerConstants.PROXY_HANDLER, new HAProxyMessageDecoder());
-            pipeline.addLast("proxyInformationHandler", createProxyHandler());
         }
 
         // Add the text line decoder which limit the max line length, don't strip the delimiter and use CRLF as delimiter
@@ -78,7 +77,7 @@ public abstract class AbstractChannelPipelineFactory<C extends SocketChannel> ex
         pipeline.addLast(HandlerConstants.CHUNK_HANDLER, new ChunkedWriteHandler());
         pipeline.addLast(HandlerConstants.TIMEOUT_HANDLER, new TimeoutHandler(timeout));
 
-        pipeline.addLast(eventExecutorGroup, HandlerConstants.CORE_HANDLER, createHandler());
+        pipeline.addLast(HandlerConstants.CORE_HANDLER, createHandler());
     }
 
     
@@ -88,10 +87,5 @@ public abstract class AbstractChannelPipelineFactory<C extends SocketChannel> ex
      * @return coreHandler
      */
     protected abstract ChannelInboundHandlerAdapter createHandler();
-
-    /**
-     * @return A handler that set HAProxy information on the underlying protocol objects.
-     */
-    protected abstract ChannelInboundHandlerAdapter createProxyHandler();
 
 }
