+++ protocols/netty/src/main/java/org/apache/james/protocols/netty/NettyServer.java
@@ -134,11 +134,6 @@ public class NettyServer extends AbstractAsyncServer {
             protected ChannelInboundHandlerAdapter createHandler() {
                 return createCoreHandler();
             }
-
-            @Override
-            protected ChannelInboundHandlerAdapter createProxyHandler() {
-                return new HAProxyMessageHandler(protocol, new ProtocolMDCContextFactory.Standard());
-            }
         };
 
     }
