+++ server/protocols/protocols-library/src/main/java/org/apache/james/protocols/lib/netty/AbstractConfigurableAsyncServer.java
@@ -452,8 +452,6 @@ public abstract class AbstractConfigurableAsyncServer extends AbstractAsyncServe
 
     protected abstract ChannelInboundHandlerAdapter createCoreHandler();
 
-    protected abstract ChannelInboundHandlerAdapter createProxyHandler();
-
     @Override
     protected AbstractChannelPipelineFactory createPipelineFactory() {
         return new AbstractSSLAwareChannelPipelineFactory<>(getTimeout(), connectionLimit, connPerIP,
@@ -463,11 +461,6 @@ public abstract class AbstractConfigurableAsyncServer extends AbstractAsyncServe
             @Override
             protected ChannelInboundHandlerAdapter createHandler() {
                 return AbstractConfigurableAsyncServer.this.createCoreHandler();
-            }
-
-            @Override
-            protected ChannelInboundHandlerAdapter createProxyHandler() {
-                return AbstractConfigurableAsyncServer.this.createProxyHandler();
             }
         };
     }
