+++ server/protocols/protocols-lmtp/src/main/java/org/apache/james/lmtpserver/netty/LMTPServer.java
@@ -33,10 +33,8 @@ import org.apache.james.protocols.lib.netty.AbstractProtocolAsyncServer;
 import org.apache.james.protocols.lmtp.LMTPConfiguration;
 import org.apache.james.protocols.netty.AbstractChannelPipelineFactory;
 import org.apache.james.protocols.netty.ChannelHandlerFactory;
-import org.apache.james.protocols.netty.HAProxyMessageHandler;
 import org.apache.james.protocols.netty.LineDelimiterBasedChannelHandlerFactory;
 import org.apache.james.protocols.smtp.SMTPProtocol;
-import org.apache.james.protocols.smtp.core.SMTPMDCContextFactory;
 import org.apache.james.smtpserver.ExtendedSMTPSession;
 import org.apache.james.smtpserver.netty.SMTPChannelInboundHandler;
 import org.slf4j.Logger;
@@ -153,17 +151,6 @@ public class LMTPServer extends AbstractProtocolAsyncServer implements LMTPServe
             }
         };
         return new SMTPChannelInboundHandler(transport, lmtpMetrics);
-    }
-
-    @Override
-    protected ChannelInboundHandlerAdapter createProxyHandler() {
-        SMTPProtocol transport = new SMTPProtocol(getProtocolHandlerChain(), lmtpConfig) {
-            @Override
-            public ProtocolSession newSession(ProtocolTransport transport) {
-                return new ExtendedSMTPSession(lmtpConfig, transport);
-            }
-        };
-        return new HAProxyMessageHandler(transport, new SMTPMDCContextFactory());
     }
 
     @Override
